+++ src/render/index.js
@@ -1,7 +1,0 @@
-import Vue from 'vue'
-import { render, helper } from 'epage-core'
-import vant from 'vant'
-
-helper.usePlugins(Vue, [vant])
-
-export default render.VueRender