+++ src/widgets/textarea/index.js
@@ -1,3 +1,2 @@
 export { default as View } from './View'
 export { default as Schema } from './Schema'
-export { default as Setting } from './Setting'